Moved misplaced classes from V7 to V4
