+++ MvvmCross.Droid.Support.V14.Preference/MvvmCross.Droid.Support.V14.Preference.csproj
@@ -110,9 +110,9 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\MvvmCross.Droid.Support.V7.Fragging\MvvmCross.Droid.Support.V7.Fragging.csproj">
-      <Project>{5B74F15E-E377-4662-9636-4233049EDDC3}</Project>
-      <Name>MvvmCross.Droid.Support.V7.Fragging</Name>
+    <ProjectReference Include="..\MvvmCross.Droid.Support.V4\MvvmCross.Droid.Support.V4.csproj">
+      <Project>{24c19002-ad02-4b88-9e57-1c254819d30e}</Project>
+      <Name>MvvmCross.Droid.Support.V4</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="..\packages\Xamarin.Android.Support.Vector.Drawable.23.2.1\build\Xamarin.Android.Support.Vector.Drawable.targets" Condition="Exists('..\packages\Xamarin.Android.Support.Vector.Drawable.23.2.1\build\Xamarin.Android.Support.Vector.Drawable.targets')" />
