+++ MvvmCross.Droid.Support.V17.Leanback/MvvmCross.Droid.Support.V17.Leanback.csproj
@@ -148,9 +148,9 @@
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
   <ItemGroup>
-    <ProjectReference Include="..\MvvmCross.Droid.Support.V7.Fragging\MvvmCross.Droid.Support.V7.Fragging.csproj">
-      <Project>{5B74F15E-E377-4662-9636-4233049EDDC3}</Project>
-      <Name>MvvmCross.Droid.Support.V7.Fragging</Name>
+    <ProjectReference Include="..\MvvmCross.Droid.Support.V4\MvvmCross.Droid.Support.V4.csproj">
+      <Project>{24c19002-ad02-4b88-9e57-1c254819d30e}</Project>
+      <Name>MvvmCross.Droid.Support.V4</Name>
     </ProjectReference>
     <ProjectReference Include="..\MvvmCross.Droid.Support.V7.RecyclerView\MvvmCross.Droid.Support.V7.RecyclerView.csproj">
       <Project>{31C39E0F-116E-4EAC-921D-A79E6AA28204}</Project>
