+++ MvvmCross.Droid.Support.V4/Caching/CachedFragmentManager.cs
