+++ MvvmCross.Droid.Support.V4/EventSource/MvxBindingFragmentAdapter.cs
@@ -14,14 +14,13 @@ using MvvmCross.Droid.Platform;
 using MvvmCross.Droid.Views;
 using MvvmCross.Core.ViewModels;
 using MvvmCross.Core.Views;
-using MvvmCross.Droid.Support.V7.Fragging.Fragments.EventSource;
 using System;
 using MvvmCross.Droid.Shared.Fragments.EventSource;
 using MvvmCross.Droid.Shared.Fragments;
 using MvvmCross.Droid.Shared;
 using MvvmCross.Droid.Shared.Attributes;
 
-namespace MvvmCross.Droid.Support.V7.Fragging.Fragments
+namespace MvvmCross.Droid.Support.V4.EventSource
 {
     public class MvxBindingFragmentAdapter
         : MvxBaseFragmentAdapter
