+++ MvvmCross.Droid.Support.V4/MvvmCross.Droid.Support.V4.csproj
@@ -75,8 +75,21 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Caching\MvxCachingFragmentActivity.cs" />
+    <Compile Include="EventSource\MvxBaseFragmentAdapter.cs" />
+    <Compile Include="EventSource\MvxBindingFragmentAdapter.cs" />
+    <Compile Include="EventSource\MvxEventSourceDialogFragment.cs" />
+    <Compile Include="EventSource\MvxEventSourceFragment.cs" />
+    <Compile Include="EventSource\MvxEventSourceFragmentActivity.cs" />
+    <Compile Include="EventSource\MvxEventSourceListFragment.cs" />
+    <Compile Include="MvxDialogFragment.cs" />
+    <Compile Include="MvxFragment.cs" />
+    <Compile Include="MvxFragmentActivity.cs" />
+    <Compile Include="MvxFragmentExtensions.cs" />
     <Compile Include="MvxFragmentPagerAdapter.cs" />
+    <Compile Include="MvxFragmentViewExtensionMethods.cs" />
     <Compile Include="MvxSwipeRefreshLayout.cs" />
+    <Compile Include="MvxTabsFragmentActivity.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="MvxFragmentStatePagerAdapter.cs" />
     <Compile Include="MvxFragmentPagerAdapter2.cs" />
@@ -87,12 +100,6 @@
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\MvvmCross.Droid.Support.V7.Fragging\MvvmCross.Droid.Support.V7.Fragging.csproj">
-      <Project>{5B74F15E-E377-4662-9636-4233049EDDC3}</Project>
-      <Name>MvvmCross.Droid.Support.V7.Fragging</Name>
-    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
 </Project>
