+++ MvvmCross.Droid.Support.V4/MvxTabsFragmentActivity.cs
@@ -17,7 +17,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace MvvmCross.Droid.Support.V7.Fragging
+namespace MvvmCross.Droid.Support.V4
 {
     public abstract class MvxTabsFragmentActivity
         : MvxFragmentActivity
