+++ MvvmCross.Droid.Support.V7.Fragging/MvvmCross.Droid.Support.V7.Fragging.csproj
@@ -77,22 +77,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AssemblyConfiguration.cs" />
-    <Compile Include="Fragments\MvxFragmentExtensions.cs" />
-    <Compile Include="Fragments\EventSource\MvxBaseFragmentAdapter.cs" />
-    <Compile Include="Fragments\MvxBindingFragmentAdapter.cs" />
     <Compile Include="Fragments\MvxSimpleLayoutInflaterHolder.cs" />
-    <Compile Include="Fragments\MvxDialogFragment.cs" />
-    <Compile Include="Fragments\EventSource\MvxEventSourceDialogFragment.cs" />
-    <Compile Include="Fragments\EventSource\MvxEventSourceFragment.cs" />
-    <Compile Include="MvxEventSourceFragmentActivity.cs" />
-    <Compile Include="Fragments\EventSource\MvxEventSourceListFragment.cs" />
-    <Compile Include="Fragments\MvxFragment.cs" />
-    <Compile Include="MvxFragmentActivity.cs" />
     <Compile Include="Fragments\MvxListFragment.cs" />
-    <Compile Include="MvxFragmentViewExtensionMethods.cs" />
-    <Compile Include="MvxTabsFragmentActivity.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Caching\MvxCachingFragmentActivity.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
@@ -106,5 +93,10 @@
   </Target>
   -->
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
-  <ItemGroup />
+  <ItemGroup>
+    <ProjectReference Include="..\MvvmCross.Droid.Support.V4\MvvmCross.Droid.Support.V4.csproj">
+      <Project>{24c19002-ad02-4b88-9e57-1c254819d30e}</Project>
+      <Name>MvvmCross.Droid.Support.V4</Name>
+    </ProjectReference>
+  </ItemGroup>
 </Project>
