+++ MvvmCross.Droid.Support.V7.Preference/MvvmCross.Droid.Support.V7.Preference.csproj
@@ -104,13 +104,13 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\MvvmCross.Droid.Support.V4\MvvmCross.Droid.Support.V4.csproj">
+      <Project>{24c19002-ad02-4b88-9e57-1c254819d30e}</Project>
+      <Name>MvvmCross.Droid.Support.V4</Name>
+    </ProjectReference>
     <ProjectReference Include="..\MvvmCross.Droid.Support.V7.AppCompat\MvvmCross.Droid.Support.V7.AppCompat.csproj">
       <Project>{911D9032-EA82-41B9-80D7-5DB84951B6E5}</Project>
       <Name>MvvmCross.Droid.Support.V7.AppCompat</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\MvvmCross.Droid.Support.V7.Fragging\MvvmCross.Droid.Support.V7.Fragging.csproj">
-      <Project>{5B74F15E-E377-4662-9636-4233049EDDC3}</Project>
-      <Name>MvvmCross.Droid.Support.V7.Fragging</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
